+++ app/models/relationship.rb
@@ -1,5 +1,7 @@
 class Relationship < ActiveRecord::Base
   serialize :properties
+
+  acts_as_paranoid
   
   belongs_to :owner, :class_name => "User"
   belongs_to :relation