+++ spec/controllers/api/oai_pmh/entities_controller_spec.rb
@@ -199,7 +199,7 @@ describe Api::OaiPmh::EntitiesController, :type => :controller do
   end
 
   it "should return 'noRecordsMatch' if the criteria do not yield any records" do
-    Entity.destroy_all
+    Entity.all.each{|r| r.really_destroy!}
     admin = User.admin
 
     get :list_identifiers, format: :xml
@@ -309,6 +309,40 @@ describe Api::OaiPmh::EntitiesController, :type => :controller do
     expect(doc.xpath('//kor:entity', ns).count).to eq(3)
 
     expect(doc.xpath('//oai:resumptionToken', ns).first). to be_nil
+  end
+
+  it 'should include deleted records' do
+    admin = User.admin
+    mona_lisa = Entity.find_by(name: 'Mona Lisa')
+    mona_lisa.destroy
+
+    get(:list_records,
+      format: :xml, 
+      api_key: admin.api_key,
+      metadataPrefix: 'kor'
+    )
+    doc = parse_xml(response.body)
+    expect(doc.xpath("//xmlns:header[@status='deleted']").count).to eq(1)
+    expect(doc.xpath("//xmlns:header[not(@status)]").count).to eq(1)
+    expect(doc.xpath("//xmlns:metadata").count).to eq(1)
+
+    get(:list_identifiers,
+      format: :xml, 
+      api_key: admin.api_key
+    )
+    doc = parse_xml(response.body)
+    expect(doc.xpath("//xmlns:header[@status='deleted']").count).to eq(1)
+    expect(doc.xpath("//xmlns:header[not(@status)]").count).to eq(1)
+    expect(doc.xpath("//xmlns:metadata").count).to eq(0)
+
+    get(:get_record,
+      format: :xml,
+      identifier: mona_lisa.uuid,
+      metadataPrefix: 'kor'
+    )
+    doc = parse_xml(response.body)
+    expect(doc.xpath("//xmlns:header[@status='deleted']").count).to eq(1)
+    expect(doc.xpath("//xmlns:metadata").count).to eq(0)
   end
 
 end